streamlit/requirements.txt
streamlit==1.28.0
pandas==2.0.3
plotly==5.17.0
pillow==10.0.1
